Update 2017: Fix relative file pathing problems
Fixed issue where when running the exe on its own, the paths to shaders
etc would be wrong. Fixed it by changing the project directory and
setting relative paths based on that.

Also removed the requirement of specifying map through argv, now it will
load a default map, which is different based on whether we are running
release or debug.